[DD-886] separated affected tests (+ decreased logging level)
